Added categories for blog, finished search options. Project is done.

+++ Blog/src/main/java/bran/packages/post/service/PostService.java
@@ -2,7 +2,10 @@ package bran.packages.post.service;
 
 import java.sql.Date;
 import java.time.LocalDate;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 import javax.transaction.Transactional;
 
@@ -11,6 +14,7 @@ import org.springframework.stereotype.Service;
 
 import bran.packages.post.dto.PostDTO;
 import bran.packages.post.entity.Post;
+import bran.packages.post.enums.CategoryEnum;
 import bran.packages.post.mapper.PostMapper;
 import bran.packages.post.repository.PostRepository;
 import bran.packages.user.entity.User;
@@ -33,23 +37,37 @@ public class PostService {
 	
 	@Transactional
 	public PostDTO update(PostDTO postDTO){
-		//TODO provere i ovde i delete
 		return postMapper.toDTO(postRepository.save(postMapper.fromDTO(postDTO)));
 	}
 	
 	@Transactional
-	public void delete(Long id){
+	public boolean delete(Long id){
+		final User user = userRepository.findByUsername(SecurityContextHolder.getContext().getAuthentication().getName()).get();
+		
+		if( user.getId() != postRepository.findByFrontId(id).getCreator().getId()) {
+			return false;
+		}
+		
 		postRepository.deleteByFrontId(id);
+		return true;
 	}
 	
 	public List<PostDTO> findPostsFromUser(String username){
 		return postMapper.toDTOList(postRepository.findByCreatorUsername(username).
-				orElseThrow( () -> new InvalidUserInfoException("bset")));
+				orElseThrow( () -> new InvalidUserInfoException("No posts from this user!")));
+	}
+	
+	public Set<String> findAllWriters(){
+		final List<PostDTO> posts = findAll();
+		final Set<String> writers = new HashSet<>();
+
+		posts.stream().forEach(post -> writers.add(post.getCreator().getUsername()));
+				
+		return writers;
 	}
 	
 	@Transactional
 	public PostDTO save(PostDTO postDTO) {
-		//TODO exeptions
 		
 		User user = userRepository.findByUsername(SecurityContextHolder.getContext().getAuthentication().getName()).get();
 		Post post = postMapper.fromDTO(postDTO);
@@ -61,6 +79,14 @@ public class PostService {
 		postRepository.save(post);
 		
 		return postMapper.toDTO(post);
+	}
+	
+	public List<PostDTO> findByCategory(CategoryEnum categoryEnum){
+		return postMapper.toDTOList(postRepository.findByCategory(categoryEnum));
+	}
+	
+	public CategoryEnum[] getCategories(){
+		return CategoryEnum.class.getEnumConstants();
 	}
 	
 	public List<PostDTO> findAll(){
